+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 * WeChatApp: 初步增加转发多客服功能
 * Record: 修复record有默认数据时,filter返回数据和默认数据混淆在一起的问题
 * Session: start方法支持session关闭后重新启动
+* Request: setServer,setQuery,setPost支持数组参数
 
 ## 0.9.20 (2015-06-18)